Define a vehicle maintenance management process
As described in the requirements, the kanban for the vehicles displays
the status of the vehicles using a color code, the colors in use are
red, yellow and green. A form for vehicle record management has been
implemented with the requested fields, there is still work to be done in
the way states are being handled for maintenance procedures and for
vehicles as well.

There are two models for this module, the 'vehicle_record.vehicle' which
describes the vehicle's properties and the
'vehicle_record.vehicle.maintenance.record0 which describes the
attributes for each entry in the record. Both models are related and
rely in each other's states to display de colors in the kanban board.